Create script to generate map background
To paste images using Pillow, we need a larger image on which we are
pasting. Furthermore, we need a grid to ensure that our tiles are
properly aligned.

This grid can further be superimposed over the structure to provide
clear squares for game purposes.

This commit creates the script which generates a background from
Blank.png tiles.