Create corridorCleaner() function for rendering
The algorithm which generates corridors draws them from the center of
shapes. Therefore, when using the vectors directly, there are large
areas of overlap between the corridors and the rooms.

corridorCleaner() solves this issue through brute force by checking if
each corridor is in any given room, then updating the variable so that
it no longer is.